Add podman in the maven image

+++ base/hack/install_utils.sh
@@ -1,24 +1,25 @@
 #!/usr/bin/env bash
 
-
 ARCH=$(uname -m)
 
 echo $ARCH
 
-# Docker
-DOCKER_VERSION=18.09.9
-if [[ ${ARCH} == 'x86_64' ]]; then
-  curl -f https://download.docker.com/linux/static/stable/x86_64/docker-$DOCKER_VERSION.tgz | tar xvz && \
-  mv docker/docker /usr/bin/ && \
-  rm -rf docker
-elif [[ ${ARCH} == 'aarch64' ]]
-then
-  curl -f https://download.docker.com/linux/static/stable/aarch64/docker-$DOCKER_VERSION.tgz | tar xvz && \
-  mv docker/docker /usr/bin/ && \
-  rm -rf docker
-else
-  echo "do not support this arch"
-  exit 1
+if [[ ${EXCLUDE_DOCKER} != '1' ]]; then
+  # Docker
+  DOCKER_VERSION=18.09.9
+  if [[ ${ARCH} == 'x86_64' ]]; then
+    curl -f https://download.docker.com/linux/static/stable/x86_64/docker-$DOCKER_VERSION.tgz | tar xvz && \
+    mv docker/docker /usr/bin/ && \
+    rm -rf docker
+  elif [[ ${ARCH} == 'aarch64' ]]
+  then
+    curl -f https://download.docker.com/linux/static/stable/aarch64/docker-$DOCKER_VERSION.tgz | tar xvz && \
+    mv docker/docker /usr/bin/ && \
+    rm -rf docker
+  else
+    echo "do not support this arch"
+    exit 1
+  fi
 fi
 
 # Helm
